New script to add a finishing pass
usage:
addfinishingpass.sh Protohead_v4_FAB_rc1-F_Cu.gbr_iso1_cnc_2_0.5mm_postprocessed.gcode 0.025

The number is the "extra depth" of the finishing pass. So if your file is coded to mill away 0.075 mm,
it will do a first pass with that amount and a second finishing pass with 0.1 mm (0.075+0.025).